# Avoid memory allocation and improve performance, part 1