--- a/src/React.Core/React.Core.csproj
+++ b/src/React.Core/React.Core.csproj
@@ -37,7 +37,7 @@
     <PackageReference Include="JavaScriptEngineSwitcher.Msie" Version="2.4.9" />
     <PackageReference Include="JSPool" Version="3.0.1" />
     <PackageReference Include="MsieJavaScriptEngine" Version="2.2.2" />
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+    <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net40' ">