… src/React.Core/GuidExtensions.cs
… src/React.Sample.Mvc4/React.Sample.Mvc4.csproj
… src/React.Sample.Mvc4/packages.config
… src/React.Sample.Owin/React.Sample.Owin.csproj
… src/React.Web.Mvc4/React.Web.Mvc4.csproj
… tests/React.Tests/Core/GuidExtensionsTests.cs
… tests/React.Tests/Core/ReactComponentTest.cs
… tests/React.Tests/Core/ReactIdGeneratorTests.cs
… tests/React.Tests/Mvc/HtmlHelperExtensionsTests.cs