Hi @Daniel15, for more than a year we have been using this library, and recently noticed problems with allocations and object hits in Large Object Heap if the component is large or a large props is serialized

Goals
1. Native implementation for IHtmlString without recopy and allocating new strings
2. Reuse everything
3. ComponentExistsChecks configuration property (default true), for disable "EnsureComponentExists" (avoid extra engine call, in production for example)


Branch with banchmarks here https://github.com/DaniilSokolyuk/React.NET/tree/bench
Note the difference between the nano seconds and microseconds

**Before, Small objects**

|              Method |     Mean |     Error |    StdDev |    Gen 0 |  Gen 1 | Allocated |
|-------------------- |---------:|----------:|----------:|---------:|-------:|----------:|
|     CreateComponent | 12.97 us | 0.0954 us | 0.0845 us |   1.8463 | 0.5035 |   10.9 KB |
|               React | 17.19 us | 0.2015 us | 0.1787 us |   7.8430 | 1.1597 |  43.81 KB |
| ReactInitJavaScript | 75.81 us | 0.5654 us | 0.5012 us | 146.9727 |      - | 602.56 KB |


**After, Small objects**

|              Method |        Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|-------------------- |------------:|----------:|----------:|-------:|-------:|----------:|
|     CreateComponent |    541.4 ns |  10.55 ns |  10.83 ns | 0.0496 | 0.0143 |     312 B |
|               React | 12,973.0 ns | 119.90 ns | 100.12 ns | 1.8005 | 0.4883 |   11320 B |
| ReactInitJavaScript | 11,401.2 ns |  72.78 ns |  64.52 ns | 0.1831 |      - |     792 B |



**Before, Large objects**

|              Method |      Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|-------------------- |----------:|----------:|----------:|---------:|---------:|--------:|-----------:|
|     CreateComponent |  35.76 us | 0.2885 us | 0.2558 us |   5.6763 |   1.5259 |       - |    34.8 KB |
|               React |  48.16 us | 0.4108 us | 0.3843 us |  24.7803 |   3.5400 |       - |  139.08 KB |
| ReactInitJavaScript | 417.16 us | 4.9953 us | 4.6726 us | 356.4453 | 142.5781 | 71.2891 | 1980.55 KB |

**After, Large objects**

|              Method |        Mean |     Error |     StdDev |   Gen 0 |  Gen 1 | Allocated |
|-------------------- |------------:|----------:|-----------:|--------:|-------:|----------:|
|     CreateComponent |    551.0 ns |  10.95 ns |   9.710 ns |  0.0496 | 0.0143 |     312 B |
|               React | 36,337.2 ns | 130.36 ns | 121.939 ns |  5.0049 | 1.3428 |   31585 B |
| ReactInitJavaScript | 51,691.7 ns | 766.71 ns | 717.185 ns | 24.8413 | 0.5493 |  104398 B |