@dustinsoftware 2018-03-24
Thanks for starting this conversation. It looks like there are a few breaking changes (like making component disposable), which could be a problem. This PR is also quite large, it would be better to split up optimization’s into smaller PRs so we can discuss each change :) 

Otherwise as is it’s going to be a bit before we can work through this one.