@DaniilSokolyuk 2018-03-24
@dustinsoftware i think disposable component is the smallest problem :)
I do not think that there are so many changes here, backward compatibility is also completely left, I do not consider it necessary to split pr