@Daniel15 2018-03-24
This is probably the largest pull request we've ever received :) Thanks for workijng on it.

There's a lot of changes that change `#if NET40` to `#if NET40 || NET451`, is it possible to split those out into a separate pull request? As @dustinsoftware said, it's much easier to review multiple small PRs instead of one giant one.