@DaniilSokolyuk 2018-03-25
@Daniel15 Thanks you, splitted, after this PR I will create another requester