@DaniilSokolyuk 2018-03-28
@Daniel15, @dustinsoftware what about this PR? I have already split this request and prepared the
[second part](https://github.com/DaniilSokolyuk/React.NET/tree/avoid-memory-allocations-part2)
We are already using these changes and significantly reduced the hits in LOH and FullGC