@dustinsoftware 2018-03-28
When you’re ready, please open some smaller pull requests so that we can
review and merge those changes individually :)

On Wed, Mar 28, 2018 at 12:08, Daniil Sokolyuk <notifications@github.com>
wrote:

> @Daniel15 <https://github.com/Daniel15>, @dustinsoftware
> <https://github.com/dustinsoftware> what about this PR? I have already
> split this request and prepared the
> second part
> <https://github.com/DaniilSokolyuk/React.NET/tree/avoid-memory-allocations-part2>
> We are already using these changes and significantly reduced the hits in
> LOH and FullGC
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/reactjs/React.NET/pull/522#issuecomment-377001291>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA5hFmwUqVYqIUKtOtV4BGc4NSQPnk2dks5ti9-OgaJpZM4S5xkt>
> .
>