@DaniilSokolyuk 2018-04-07
@dustinsoftware, yes you right, but previous ID generating is ugly, we are allocate GUID, after allocate byte array and 1-3 string...  my implementation used 1 string