@dustinsoftware 2018-04-07
> previous ID generating is ugly

There may be consuming code that depends on the current ID format, so this isn't a good reason to change it. We unfortunately don't know how consumers are using this project and want to avoid surprises between version updates.