@DaniilSokolyuk 2018-04-08
@damiangreen, we can announce this change or publish major version