@dustinsoftware 2018-04-08
That’s not the point I was trying to make... I am hesitant to merge in changes for minor improvements that are technically breaking changes. 

The discussion on this PR is getting too long, please open a new one for each individual change (such as ones the bulleted list above) and we can discuss them one at a time :)